check

+++ backend/src/main/java/com/backend/backend/DAO/Pets/PetRepository.java
@@ -1,5 +1,7 @@
 package com.backend.backend.DAO.Pets;
 
+import java.math.BigInteger;
+
 import com.backend.backend.Entity.Pet;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -7,7 +9,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface PetRepository extends MongoRepository<Pet, Integer>{
+public interface PetRepository extends MongoRepository<Pet, BigInteger>{
 
     
 }
